[TPDE,X64] Use negative frame offsets
This reduces the number of cases where negative/positive frame offsets
need to be distinguished.

+++ tpde-llvm/src/arm64/LLVMCompilerArm64.cpp
@@ -213,7 +213,7 @@ void LLVMCompilerArm64::load_address_of_var_reference(
   const auto &info = variable_refs[ap.assignment->var_ref_custom_idx];
   if (info.alloca) {
     // default handling from CompilerA64
-    assert(static_cast<i32>(info.alloca_frame_off) >= 0);
+    assert(info.alloca_frame_off >= 0);
     // per-default, variable references are only used by
     // allocas
     if (!ASMIF(ADDxi, dst, DA_GP(29), info.alloca_frame_off)) {
